+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/wodclipse/core/woo/WooModel.java
@@ -11,7 +11,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -60,10 +59,6 @@ public class WooModel {
     catch (Throwable e) {
       throw new RuntimeException(e);
     }
-  }
-
-  public WooModel(final URL url) {
-    // TODO: Fix me
   }
 
   public WooModel(final String contents) throws WooModelException {